// backend/src/domains/portfolio-projects/data-transfer-objects/update-project-response.dto.ts
export class ProjectMedia {
  id: number;
  filepath: string;
  size: number;
  type: string;
  filename: string;
}

export class UpdateProjectResponse {
  id: number;
  title: string;
  category: string;
  text: string;
  imageMediaId: number;
  background: string;
  link: string;
  medias: ProjectMedia[];
}